April Showers Update and Variables Change
Just_Light_Accent1 and Just_Light_Accent2 are now called Just_White_Accent1 and Just_White_Accent2 to play more on the basis of being the opposite of the theme's name, Just_Black.  This changes it so that only the Dark and Light variables make direct allusions to Dark Theme and Light Theme, respectively.